+++ lib/td_dq_web/controllers/rule_filter_controller.ex
@@ -1,10 +1,9 @@
 defmodule TdDqWeb.RuleFilterController do
-  require Logger
   use TdDqWeb, :controller
-  use PhoenixSwagger
 
   alias TdDq.Rules.Search
-  alias TdDqWeb.SwaggerDefinitions
+
+  require Logger
 
   action_fallback(TdDqWeb.FallbackController)