+++ lib/td_dq_web/controllers/search_controller.ex
@@ -1,6 +1,5 @@
 defmodule TdDqWeb.SearchController do
   use TdDqWeb, :controller
-  use PhoenixSwagger
 
   import Canada, only: [can?: 2]